+++ packages/eyes-common/lib/config/IosScreenOrientation.js
@@ -1,0 +1,15 @@
+'use strict'
+
+/**
+ * @readonly
+ * @enum {string}
+ */
+const IosScreenOrientation = {
+  PORTRAIT: 'portrait',
+  UPSIDE_DOWN: 'upsideDown',
+  LANDSCAPE_LEFT: 'landscapeLeft',
+  LANDSCAPE_RIGHT: 'landscapeRight',
+}
+
+Object.freeze(IosScreenOrientation)
+exports.IosScreenOrientation = IosScreenOrientation